refactor(board): create board service folder to organize service file
