+++ src/app/state/board/board.effects.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { of } from 'rxjs';
 import { catchError, map, mergeMap } from 'rxjs/operators';
-import { BoardService } from 'src/app/service/board.service';
+import { BoardService } from 'src/app/service/board/board.service';
 import {
   fetchBoard,
   fetchBoardError,
